# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для English Bot - системы изучения английского языка с использованием SRS (Spaced Repetition System)",
        "title": "English Bot API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8184",
    "basePath": "/",
    "paths": {
        "/app/admin": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает состояние circuit breaker и информацию об администраторе (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Получить данные админ-панели",
                "responses": {
                    "200": {
                        "description": "Данные админ-панели",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен (требуются права администратора)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Метод не разрешен",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/app/admin/circuit/reset": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Сбрасывает состояние circuit breaker (только для администраторов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Сбросить circuit breaker",
                "responses": {
                    "200": {
                        "description": "Успешный сброс",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен (требуются права администратора)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/app/admin/training/{word}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение (GET) или удаление (POST) тренировочных карточек по слову. Путь: /app/admin/training/{word} или /app/admin/training/{word}/delete или /app/admin/training/delete_all",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Управление тренировочными карточками",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Английское слово",
                        "name": "word",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Действие: delete или delete_all",
                        "name": "action",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Английское слово (для GET запроса)",
                        "name": "word",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные карточек или результат удаления",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен (требуются права администратора)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer JWT token authentication. В Swagger UI введите только токен (без \"Bearer \"), префикс будет добавлен автоматически.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}